started corpus generation code

+++ src/simplexai/experiments/corpus_generation.py
@@ -47,7 +47,7 @@ def main():
         prototype_df.append(prototype)
 
     # Binarize diagnosis features using the occurance probability as threshold
-    prototype_df = pd.DataFrame.from_dict(prototype_df, orient='columns')
+    prototype_df = pd.DataFrame.from_dict(prototype_df, orient='columns').astype('float')
     print(prototype_df)
     prototype_df.iloc[:, -23:] = prototype_df.iloc[:, -23:] >= diagnoses_probs
     print(prototype_df)
